tools/traffic-generator/pyproject.toml
[project]
name = "octo-traffic-generator"
version = "1.0.0"
description = "Realistic synthetic traffic for octo-apm-demo — drone.octodemo.cloud + backend.octodemo.cloud"
requires-python = ">=3.11"
dependencies = [
    "httpx[http2]>=0.27",
    "structlog>=24.1",
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp-proto-http>=1.24",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "numpy>=1.26",  # realistic distributions
    "rich>=13.7",   # operator-facing progress output
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "respx>=0.21"]

[project.scripts]
octo-traffic = "octo_traffic.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
